fix "Edit me" button URLs.
The configuration needed to point to the cloudmaven repository for the
"Edit me" buttons to work properly. Also, for the markdown files to link
correctly, I needed to update the github_editme_path portion of
page.html to match the latest version from the theme:
https://github.com/tomjohnson1492/documentation-theme-jekyll/blob/e4ed2d864d4c2a28ae232ba7a6c9485feed234ae/_layouts/page.html

Also, update the version of JSON used in Gemfile.lock.  I was getting
compilation errors when trying to test changes locally with `bundle exec
jekyll serve` per the instructions at
https://help.github.com/articles/setting-up-your-github-pages-site-locally-with-jekyll/